initial refactors to split core into renderer and vue-renderer
