+++ packages/core/package.json
@@ -9,33 +9,17 @@
   "main": "dist/core.js",
   "dependencies": {
     "@nuxt/common": "^2.2.0",
+    "@nuxt/server": "^2.2.0",
+    "@nuxt/vue-renderer": "^2.2.0",
     "@nuxtjs/devalue": "^1.0.1",
     "@nuxtjs/opencollective": "^0.1.0",
-    "@nuxtjs/youch": "^4.2.3",
-    "chalk": "^2.4.1",
-    "compression": "^1.7.3",
-    "connect": "^3.6.6",
     "consola": "^1.4.4",
     "debug": "^4.1.0",
     "esm": "^3.0.84",
-    "etag": "^1.8.1",
-    "fresh": "^0.5.2",
     "fs-extra": "^7.0.0",
     "hash-sum": "^1.0.2",
-    "ip": "^1.1.5",
-    "launch-editor-middleware": "^2.2.1",
     "lodash": "^4.17.11",
-    "lru-cache": "^4.1.3",
-    "serve-static": "^1.13.2",
-    "server-destroy": "^1.0.1",
-    "std-env": "^2.0.2",
-    "vue": "^2.5.17",
-    "vue-meta": "^1.5.5",
-    "vue-no-ssr": "^1.0.0",
-    "vue-router": "^3.0.1",
-    "vue-server-renderer": "^2.5.17",
-    "vue-template-compiler": "^2.5.17",
-    "vuex": "^3.0.1"
+    "std-env": "^2.0.2"
   },
   "publishConfig": {
     "access": "public"
