+++ packages/core/src/index.js
@@ -1,3 +1,3 @@
 export { default as Module } from './module'
 export { default as Nuxt } from './nuxt'
-export { default as Renderer } from './renderer'
+export { default as Resolver } from './resolver'
