+++ packages/server/package.js
@@ -1,0 +1,3 @@
+export default {
+  build: true
+}
