+++ packages/server/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@nuxt/server",
+  "version": "2.2.0",
+  "repository": "nuxt/nuxt.js",
+  "license": "MIT",
+  "files": [
+    "dist"
+  ],
+  "main": "dist/server.js",
+  "dependencies": {
+    "@nuxt/common": "^2.2.0",
+    "@nuxtjs/youch": "^4.2.3",
+    "chalk": "^2.4.1",
+    "compression": "^1.7.3",
+    "connect": "^3.6.6",
+    "consola": "^1.4.4",
+    "etag": "^1.8.1",
+    "fresh": "^0.5.2",
+    "fs-extra": "^7.0.0",
+    "ip": "^1.1.5",
+    "launch-editor-middleware": "^2.2.1",
+    "serve-static": "^1.13.2",
+    "server-destroy": "^1.0.1"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
