+++ packages/server/src/index.js
@@ -1,0 +1,1 @@
+export { default as Server } from './server'
