+++ packages/vue-renderer/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@nuxt/vue-renderer",
+  "version": "2.2.0",
+  "repository": "nuxt/nuxt.js",
+  "license": "MIT",
+  "files": [
+    "dist"
+  ],
+  "main": "dist/vue-renderer.js",
+  "dependencies": {
+    "@nuxt/common": "^2.2.0",
+    "@nuxtjs/devalue": "^1.0.1",
+    "consola": "^1.4.4",
+    "fs-extra": "^7.0.0",
+    "lru-cache": "^4.1.3",
+    "vue": "^2.5.17",
+    "vue-meta": "^1.5.5",
+    "vue-no-ssr": "^1.0.0",
+    "vue-router": "^3.0.1",
+    "vue-server-renderer": "^2.5.17",
+    "vue-template-compiler": "^2.5.17",
+    "vuex": "^3.0.1"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
