+++ packages/vue-renderer/src/index.js
@@ -1,0 +1,1 @@
+export { default as VueRenderer } from './renderer'
